Modify design of pages
I have used the knoweldge I have gained about boxes to center the titles
and images of the 4 webpages currently present in this directory.

For the BBQ recipe I have tried to place the 4 divs on the same line. For
now, I am happy with the result, but it is obvious to me that there must
exist better ways to do that, such as using tables. I imagine that I will
get there at some point in my learning journey. I also realize that my
setups at the moment is not dynamic and that it will likely break on
different devices. Something to learn to fix in the future.